+++ packages/cordova/src/android/cordova/ads/RewardedInterstitial.kt
@@ -2,14 +2,15 @@ package admob.plus.cordova.ads
 
 import admob.plus.cordova.Events
 import admob.plus.cordova.ExecuteContext
-import com.google.android.gms.ads.AdError
-import com.google.android.gms.ads.FullScreenContentCallback
-import com.google.android.gms.ads.LoadAdError
-import com.google.android.gms.ads.rewarded.RewardItem
-import com.google.android.gms.ads.rewardedinterstitial.RewardedInterstitialAd
-import com.google.android.gms.ads.rewardedinterstitial.RewardedInterstitialAdLoadCallback
+import com.google.android.libraries.ads.mobile.sdk.common.AdLoadCallback
+import com.google.android.libraries.ads.mobile.sdk.common.FullScreenContentError
+import com.google.android.libraries.ads.mobile.sdk.common.LoadAdError
+import com.google.android.libraries.ads.mobile.sdk.rewarded.RewardItem
+import com.google.android.libraries.ads.mobile.sdk.rewardedinterstitial.RewardedInterstitialAd
+import com.google.android.libraries.ads.mobile.sdk.rewardedinterstitial.RewardedInterstitialAdEventCallback
 
 class RewardedInterstitial(ctx: ExecuteContext) : AdBase(ctx) {
+    @Volatile
     private var mAd: RewardedInterstitialAd? = null
     override fun onDestroy() {
         clear()
@@ -19,10 +20,8 @@ class RewardedInterstitial(ctx: ExecuteContext) : AdBase(ctx) {
     override fun load(ctx: ExecuteContext) {
         clear()
         RewardedInterstitialAd.load(
-            plugin.activity,
-            adUnitId,
             adRequest,
-            object : RewardedInterstitialAdLoadCallback() {
+            object : AdLoadCallback<RewardedInterstitialAd> {
                 override fun onAdFailedToLoad(loadAdError: LoadAdError) {
                     mAd = null
                     emit(Events.AD_LOAD_FAIL, loadAdError)
@@ -35,12 +34,15 @@ class RewardedInterstitial(ctx: ExecuteContext) : AdBase(ctx) {
                     if (ssv != null) {
                         mAd!!.setServerSideVerificationOptions(ssv)
                     }
-                    mAd!!.fullScreenContentCallback = object : FullScreenContentCallback() {
+                    mAd!!.adEventCallback = object : RewardedInterstitialAdEventCallback {
                         override fun onAdDismissedFullScreenContent() {
+                            rewardedAd.destroy()
                             emit(Events.AD_DISMISS)
                         }
 
-                        override fun onAdFailedToShowFullScreenContent(adError: AdError) {
+                        override fun onAdFailedToShowFullScreenContent(adError: FullScreenContentError) {
+                            mAd = null
+                            rewardedAd.destroy()
                             emit(Events.AD_SHOW_FAIL, adError)
                         }
 
@@ -51,6 +53,10 @@ class RewardedInterstitial(ctx: ExecuteContext) : AdBase(ctx) {
 
                         override fun onAdImpression() {
                             emit(Events.AD_IMPRESSION)
+                        }
+
+                        override fun onAdClicked() {
+                            emit(Events.AD_CLICK)
                         }
                     }
                     emit(Events.AD_LOAD)
@@ -74,8 +80,7 @@ class RewardedInterstitial(ctx: ExecuteContext) : AdBase(ctx) {
     }
 
     private fun clear() {
-        if (mAd != null) {
-            mAd = null
-        }
+        mAd?.destroy()
+        mAd = null
     }
 }